Handle errors thrown from benchcomp column exprs (#3145)
Prior to this commit, errors thrown when evaluating the text of a
benchcomp extra column would crash benchcomp. This could happen, for
example, if a column tries to compare an old variant with a new one, but
no data for the old variant exists, as seen in this run:

https://github.com/model-checking/kani/actions/runs/8700040930/job/23859607740

Forcing the user to do error handling in the column text would make the
text even more unwieldy than it already is, so this commit makes the
column text evaluate to **<ERROR>** if an exception is raised during
evaluation.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.